--- G2-F/fooGL.py
@@ -1,447 +1,0 @@
-import sys
-import math
-
-from OpenGL import GL
-from OpenGL import GLU
-from OpenGL import GLUT
-
-from PyQt5.QtCore import pyqtSignal, QPoint, QSize, Qt
-from PyQt5.QtGui import QColor
-from PyQt5.QtWidgets import (QApplication, QHBoxLayout, QOpenGLWidget, QSlider, QWidget)
-
-
-class GLWidget(QOpenGLWidget):
-    xRotationChanged = pyqtSignal(int)
-    yRotationChanged = pyqtSignal(int)
-    zRotationChanged = pyqtSignal(int)
-
-    def __init__(self, parent=None):
-        super(GLWidget, self).__init__(parent)
-
-        self.object = 0
-        self.xRot = 0
-        self.yRot = 0
-        self.zRot = 0
-
-        self.lastPos = QPoint()
-
-        self.trolltechGreen = QColor.fromCmykF(0.40, 0.0, 1.0, 0.0)
-        self.trolltechPurple = QColor.fromCmykF(0.39, 0.39, 0.0, 0.0)
-
-    def minimumSizeHint(self):
-        return QSize(50, 50)
-
-    def sizeHint(self):
-        return QSize(400, 400)
-
-    def setXRotation(self, angle):
-        angle = self.normalizeAngle(angle)
-        if angle != self.xRot:
-            self.xRot = angle
-            self.xRotationChanged.emit(angle)
-            self.update()
-
-    def setYRotation(self, angle):
-        angle = self.normalizeAngle(angle)
-        if angle != self.yRot:
-            self.yRot = angle
-            self.yRotationChanged.emit(angle)
-            self.update()
-
-    def setZRotation(self, angle):
-        angle = self.normalizeAngle(angle)
-        if angle != self.zRot:
-            self.zRot = angle
-            self.zRotationChanged.emit(angle)
-            self.update()
-
-    def initializeGL(self):
-        #self.gl = self.context().versionFunctions()
-        self.gl = GL
-        #self.gl.initializeOpenGLFunctions()
-
-        self.setClearColor(self.trolltechPurple.darker())
-        self.object = self.makeObject()
-        self.gl.glShadeModel(self.gl.GL_FLAT)
-        self.gl.glEnable(self.gl.GL_DEPTH_TEST)
-        self.gl.glEnable(self.gl.GL_CULL_FACE)
-
-    def paintGL(self):
-        self.gl.glClear(
-                self.gl.GL_COLOR_BUFFER_BIT | self.gl.GL_DEPTH_BUFFER_BIT)
-        self.gl.glLoadIdentity()
-        self.gl.glTranslated(0.0, 0.0, -10.0)
-        self.gl.glRotated(self.xRot / 16.0, 1.0, 0.0, 0.0)
-        self.gl.glRotated(self.yRot / 16.0, 0.0, 1.0, 0.0)
-        self.gl.glRotated(self.zRot / 16.0, 0.0, 0.0, 1.0)
-        self.gl.glCallList(self.object)
-
-    def resizeGL(self, width, height):
-        side = min(width, height)
-        if side < 0:
-            return
-
-        self.gl.glViewport((width - side) // 2, (height - side) // 2, side,
-                side)
-
-        self.gl.glMatrixMode(self.gl.GL_PROJECTION)
-        self.gl.glLoadIdentity()
-        self.gl.glOrtho(-0.5, +0.5, +0.5, -0.5, 4.0, 15.0)
-        self.gl.glMatrixMode(self.gl.GL_MODELVIEW)
-
-    def mousePressEvent(self, event):
-        self.lastPos = event.pos()
-
-    def mouseMoveEvent(self, event):
-        dx = event.x() - self.lastPos.x()
-        dy = event.y() - self.lastPos.y()
-
-        if event.buttons() & Qt.LeftButton:
-            self.setXRotation(self.xRot + 8 * dy)
-            self.setYRotation(self.yRot + 8 * dx)
-        elif event.buttons() & Qt.RightButton:
-            self.setXRotation(self.xRot + 8 * dy)
-            self.setZRotation(self.zRot + 8 * dx)
-
-        self.lastPos = event.pos()
-
-    def makeObject(self):
-        genList = self.gl.glGenLists(1)
-        self.gl.glNewList(genList, self.gl.GL_COMPILE)
-
-        self.gl.glBegin(self.gl.GL_QUADS)
-
-        x1 = +0.06
-        y1 = -0.14
-        x2 = +0.14
-        y2 = -0.06
-        x3 = +0.08
-        y3 = +0.00
-        x4 = +0.30
-        y4 = +0.22
-
-        self.quad(x1, y1, x2, y2, y2, x2, y1, x1)
-        self.quad(x3, y3, x4, y4, y4, x4, y3, x3)
-
-        self.extrude(x1, y1, x2, y2)
-        self.extrude(x2, y2, y2, x2)
-        self.extrude(y2, x2, y1, x1)
-        self.extrude(y1, x1, x1, y1)
-        self.extrude(x3, y3, x4, y4)
-        self.extrude(x4, y4, y4, x4)
-        self.extrude(y4, x4, y3, x3)
-
-        NumSectors = 50
-
-        for i in range(NumSectors):
-            angle1 = (i * 2 * math.pi) / NumSectors
-            x5 = 0.30 * math.sin(angle1)
-            y5 = 0.30 * math.cos(angle1)
-            x6 = 0.20 * math.sin(angle1)
-            y6 = 0.20 * math.cos(angle1)
-
-            angle2 = ((i + 1) * 2 * math.pi) / NumSectors
-            x7 = 0.20 * math.sin(angle2)
-            y7 = 0.20 * math.cos(angle2)
-            x8 = 0.30 * math.sin(angle2)
-            y8 = 0.30 * math.cos(angle2)
-
-            self.quad(x5, y5, x6, y6, x7, y7, x8, y8)
-
-            self.extrude(x6, y6, x7, y7)
-            self.extrude(x8, y8, x5, y5)
-
-        self.gl.glEnd()
-        self.gl.glEndList()
-
-        return genList
-
-    def quad(self, x1, y1, x2, y2, x3, y3, x4, y4):
-        self.setColor(self.trolltechGreen)
-
-        self.gl.glVertex3d(x1, y1, -0.05)
-        self.gl.glVertex3d(x2, y2, -0.05)
-        self.gl.glVertex3d(x3, y3, -0.05)
-        self.gl.glVertex3d(x4, y4, -0.05)
-
-        self.gl.glVertex3d(x4, y4, +0.05)
-        self.gl.glVertex3d(x3, y3, +0.05)
-        self.gl.glVertex3d(x2, y2, +0.05)
-        self.gl.glVertex3d(x1, y1, +0.05)
-
-    def extrude(self, x1, y1, x2, y2):
-        self.setColor(self.trolltechGreen.darker(250 + int(100 * x1)))
-
-        self.gl.glVertex3d(x1, y1, +0.05)
-        self.gl.glVertex3d(x2, y2, +0.05)
-        self.gl.glVertex3d(x2, y2, -0.05)
-        self.gl.glVertex3d(x1, y1, -0.05)
-
-    def normalizeAngle(self, angle):
-        while angle < 0:
-            angle += 360 * 16
-        while angle > 360 * 16:
-            angle -= 360 * 16
-        return angle
-
-    def setClearColor(self, c):
-        self.gl.glClearColor(c.redF(), c.greenF(), c.blueF(), c.alphaF())
-
-    def setColor(self, c):
-        self.gl.glColor4f(c.redF(), c.greenF(), c.blueF(), c.alphaF())
-
-
-# -*- coding: utf-8 -*-
-
-# Form implementation generated from reading ui file 'mainwindow.ui'
-#
-# Created by: PyQt5 UI code generator 5.9.1
-#
-# WARNING! All changes made in this file will be lost!
-
-from PyQt5 import QtCore, QtGui, QtWidgets
-
-class Ui_MainWindow(object):
-    def setupUi(self, MainWindow):
-        MainWindow.setObjectName("MainWindow")
-        MainWindow.setWindowModality(QtCore.Qt.WindowModal)
-        MainWindow.resize(1000, 600)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Preferred, QtWidgets.QSizePolicy.Preferred)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(MainWindow.sizePolicy().hasHeightForWidth())
-        MainWindow.setSizePolicy(sizePolicy)
-        self.centralWidget = QtWidgets.QWidget(MainWindow)
-        self.centralWidget.setObjectName("centralWidget")
-        self.layoutWidget = QtWidgets.QWidget(self.centralWidget)
-        self.layoutWidget.setGeometry(QtCore.QRect(0, 0, 1001, 411))
-        self.layoutWidget.setObjectName("layoutWidget")
-        self.gridLayout = QtWidgets.QGridLayout(self.layoutWidget)
-        self.gridLayout.setContentsMargins(11, 11, 11, 11)
-        self.gridLayout.setSpacing(6)
-        self.gridLayout.setObjectName("gridLayout")
-        spacerItem = QtWidgets.QSpacerItem(20, 20, QtWidgets.QSizePolicy.Minimum, QtWidgets.QSizePolicy.Fixed)
-        self.gridLayout.addItem(spacerItem, 0, 0, 1, 1)
-        self.gridLayout_2 = QtWidgets.QGridLayout()
-        self.gridLayout_2.setSpacing(6)
-        self.gridLayout_2.setObjectName("gridLayout_2")
-
-        self.openGLWidget = GLWidget(self.layoutWidget)
-        self.openGLWidget.setEnabled(True)
-        self.openGLWidget.setObjectName("openGLWidget")
-
-        self.gridLayout_2.addWidget(self.openGLWidget, 0, 1, 1, 1)
-        spacerItem1 = QtWidgets.QSpacerItem(20, 20, QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Minimum)
-        self.gridLayout_2.addItem(spacerItem1, 0, 0, 1, 1)
-        spacerItem2 = QtWidgets.QSpacerItem(20, 20, QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Minimum)
-        self.gridLayout_2.addItem(spacerItem2, 0, 2, 1, 1)
-        spacerItem3 = QtWidgets.QSpacerItem(20, 20, QtWidgets.QSizePolicy.Minimum, QtWidgets.QSizePolicy.Fixed)
-        self.gridLayout_2.addItem(spacerItem3, 1, 1, 1, 1)
-        self.gridLayout.addLayout(self.gridLayout_2, 1, 0, 1, 1)
-        self.widget = QtWidgets.QWidget(self.centralWidget)
-        self.widget.setGeometry(QtCore.QRect(0, 410, 1001, 141))
-        self.widget.setObjectName("widget")
-        self.gridLayout_3 = QtWidgets.QGridLayout(self.widget)
-        self.gridLayout_3.setContentsMargins(11, 11, 11, 11)
-        self.gridLayout_3.setSpacing(6)
-        self.gridLayout_3.setObjectName("gridLayout_3")
-        spacerItem4 = QtWidgets.QSpacerItem(20, 20, QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Minimum)
-        self.gridLayout_3.addItem(spacerItem4, 0, 0, 1, 1)
-        self.gridLayout_4 = QtWidgets.QGridLayout()
-        self.gridLayout_4.setSpacing(6)
-        self.gridLayout_4.setObjectName("gridLayout_4")
-        spacerItem5 = QtWidgets.QSpacerItem(20, 35, QtWidgets.QSizePolicy.Minimum, QtWidgets.QSizePolicy.Fixed)
-        self.gridLayout_4.addItem(spacerItem5, 0, 0, 1, 1)
-        self.comboBox = QtWidgets.QComboBox(self.widget)
-        self.comboBox.setObjectName("comboBox")
-        self.gridLayout_4.addWidget(self.comboBox, 1, 0, 1, 1)
-        self.label_4 = QtWidgets.QLabel(self.widget)
-        self.label_4.setObjectName("label_4")
-        self.gridLayout_4.addWidget(self.label_4, 2, 0, 1, 1)
-        spacerItem6 = QtWidgets.QSpacerItem(20, 35, QtWidgets.QSizePolicy.Minimum, QtWidgets.QSizePolicy.Fixed)
-        self.gridLayout_4.addItem(spacerItem6, 3, 0, 1, 1)
-        self.gridLayout_3.addLayout(self.gridLayout_4, 0, 1, 1, 1)
-        self.line_3 = QtWidgets.QFrame(self.widget)
-        self.line_3.setFrameShape(QtWidgets.QFrame.VLine)
-        self.line_3.setFrameShadow(QtWidgets.QFrame.Sunken)
-        self.line_3.setObjectName("line_3")
-        self.gridLayout_3.addWidget(self.line_3, 0, 2, 1, 1)
-        self.gridLayout_5 = QtWidgets.QGridLayout()
-        self.gridLayout_5.setSpacing(6)
-        self.gridLayout_5.setObjectName("gridLayout_5")
-        self.dial = QtWidgets.QDial(self.widget)
-        self.dial.setObjectName("dial")
-        self.gridLayout_5.addWidget(self.dial, 0, 0, 1, 1)
-        self.label = QtWidgets.QLabel(self.widget)
-        self.label.setObjectName("label")
-        self.gridLayout_5.addWidget(self.label, 1, 0, 1, 1)
-        self.gridLayout_3.addLayout(self.gridLayout_5, 0, 3, 1, 1)
-        self.gridLayout_6 = QtWidgets.QGridLayout()
-        self.gridLayout_6.setSpacing(6)
-        self.gridLayout_6.setObjectName("gridLayout_6")
-        self.dial_2 = QtWidgets.QDial(self.widget)
-        self.dial_2.setObjectName("dial_2")
-        self.gridLayout_6.addWidget(self.dial_2, 0, 0, 1, 1)
-        self.label_2 = QtWidgets.QLabel(self.widget)
-        self.label_2.setObjectName("label_2")
-        self.gridLayout_6.addWidget(self.label_2, 1, 0, 1, 1)
-        self.gridLayout_3.addLayout(self.gridLayout_6, 0, 4, 1, 1)
-        self.gridLayout_7 = QtWidgets.QGridLayout()
-        self.gridLayout_7.setSpacing(6)
-        self.gridLayout_7.setObjectName("gridLayout_7")
-        self.dial_3 = QtWidgets.QDial(self.widget)
-        self.dial_3.setObjectName("dial_3")
-        self.gridLayout_7.addWidget(self.dial_3, 0, 0, 1, 1)
-        self.label_3 = QtWidgets.QLabel(self.widget)
-        self.label_3.setObjectName("label_3")
-        self.gridLayout_7.addWidget(self.label_3, 1, 0, 1, 1)
-        self.gridLayout_3.addLayout(self.gridLayout_7, 0, 5, 1, 1)
-        self.line = QtWidgets.QFrame(self.widget)
-        self.line.setFrameShape(QtWidgets.QFrame.VLine)
-        self.line.setFrameShadow(QtWidgets.QFrame.Sunken)
-        self.line.setObjectName("line")
-        self.gridLayout_3.addWidget(self.line, 0, 6, 1, 1)
-        self.gridLayout_8 = QtWidgets.QGridLayout()
-        self.gridLayout_8.setSpacing(6)
-        self.gridLayout_8.setObjectName("gridLayout_8")
-        spacerItem7 = QtWidgets.QSpacerItem(20, 35, QtWidgets.QSizePolicy.Minimum, QtWidgets.QSizePolicy.Fixed)
-        self.gridLayout_8.addItem(spacerItem7, 0, 0, 1, 1)
-        self.pushButton = QtWidgets.QPushButton(self.widget)
-        self.pushButton.setObjectName("pushButton")
-        self.gridLayout_8.addWidget(self.pushButton, 1, 0, 1, 1)
-        self.pushButton_2 = QtWidgets.QPushButton(self.widget)
-        self.pushButton_2.setObjectName("pushButton_2")
-        self.gridLayout_8.addWidget(self.pushButton_2, 2, 0, 1, 1)
-        spacerItem8 = QtWidgets.QSpacerItem(20, 35, QtWidgets.QSizePolicy.Minimum, QtWidgets.QSizePolicy.Fixed)
-        self.gridLayout_8.addItem(spacerItem8, 3, 0, 1, 1)
-        self.gridLayout_3.addLayout(self.gridLayout_8, 0, 7, 1, 1)
-        self.line_2 = QtWidgets.QFrame(self.widget)
-        self.line_2.setFrameShape(QtWidgets.QFrame.VLine)
-        self.line_2.setFrameShadow(QtWidgets.QFrame.Sunken)
-        self.line_2.setObjectName("line_2")
-        self.gridLayout_3.addWidget(self.line_2, 0, 8, 1, 1)
-        self.horizontalLayout = QtWidgets.QHBoxLayout()
-        self.horizontalLayout.setSpacing(6)
-        self.horizontalLayout.setObjectName("horizontalLayout")
-        self.gridLayout_9 = QtWidgets.QGridLayout()
-        self.gridLayout_9.setSpacing(6)
-        self.gridLayout_9.setObjectName("gridLayout_9")
-        spacerItem9 = QtWidgets.QSpacerItem(20, 35, QtWidgets.QSizePolicy.Minimum, QtWidgets.QSizePolicy.Fixed)
-        self.gridLayout_9.addItem(spacerItem9, 0, 0, 1, 1)
-        self.lcdNumber = QtWidgets.QLCDNumber(self.widget)
-        self.lcdNumber.setObjectName("lcdNumber")
-        self.gridLayout_9.addWidget(self.lcdNumber, 1, 0, 1, 1)
-        self.label_5 = QtWidgets.QLabel(self.widget)
-        self.label_5.setObjectName("label_5")
-        self.gridLayout_9.addWidget(self.label_5, 2, 0, 1, 1)
-        spacerItem10 = QtWidgets.QSpacerItem(20, 35, QtWidgets.QSizePolicy.Minimum, QtWidgets.QSizePolicy.Fixed)
-        self.gridLayout_9.addItem(spacerItem10, 3, 0, 1, 1)
-        self.horizontalLayout.addLayout(self.gridLayout_9)
-        self.verticalSlider = QtWidgets.QSlider(self.widget)
-        self.verticalSlider.setOrientation(QtCore.Qt.Vertical)
-        self.verticalSlider.setObjectName("verticalSlider")
-        self.horizontalLayout.addWidget(self.verticalSlider)
-        self.gridLayout_3.addLayout(self.horizontalLayout, 0, 9, 1, 1)
-        spacerItem11 = QtWidgets.QSpacerItem(20, 20, QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Minimum)
-        self.gridLayout_3.addItem(spacerItem11, 0, 10, 1, 1)
-        MainWindow.setCentralWidget(self.centralWidget)
-        self.menuBar = QtWidgets.QMenuBar(MainWindow)
-        self.menuBar.setGeometry(QtCore.QRect(0, 0, 1000, 25))
-        self.menuBar.setObjectName("menuBar")
-        self.menuFile = QtWidgets.QMenu(self.menuBar)
-        self.menuFile.setObjectName("menuFile")
-        self.menuLoad = QtWidgets.QMenu(self.menuFile)
-        self.menuLoad.setObjectName("menuLoad")
-        self.menuSave = QtWidgets.QMenu(self.menuFile)
-        self.menuSave.setObjectName("menuSave")
-        self.menuEdit = QtWidgets.QMenu(self.menuBar)
-        self.menuEdit.setObjectName("menuEdit")
-        self.menuTools = QtWidgets.QMenu(self.menuBar)
-        self.menuTools.setObjectName("menuTools")
-        self.menuHelp = QtWidgets.QMenu(self.menuBar)
-        self.menuHelp.setObjectName("menuHelp")
-        MainWindow.setMenuBar(self.menuBar)
-        self.statusBar = QtWidgets.QStatusBar(MainWindow)
-        self.statusBar.setObjectName("statusBar")
-        MainWindow.setStatusBar(self.statusBar)
-        self.actionExit = QtWidgets.QAction(MainWindow)
-        self.actionExit.setObjectName("actionExit")
-        self.actionObject = QtWidgets.QAction(MainWindow)
-        self.actionObject.setObjectName("actionObject")
-        self.actionModel = QtWidgets.QAction(MainWindow)
-        self.actionModel.setObjectName("actionModel")
-        self.actionObject_2 = QtWidgets.QAction(MainWindow)
-        self.actionObject_2.setObjectName("actionObject_2")
-        self.actionModel_2 = QtWidgets.QAction(MainWindow)
-        self.actionModel_2.setObjectName("actionModel_2")
-        self.actionPredefined_Forms = QtWidgets.QAction(MainWindow)
-        self.actionPredefined_Forms.setObjectName("actionPredefined_Forms")
-        self.actionPredefined_Scenes = QtWidgets.QAction(MainWindow)
-        self.actionPredefined_Scenes.setObjectName("actionPredefined_Scenes")
-        self.actionObject_3 = QtWidgets.QAction(MainWindow)
-        self.actionObject_3.setObjectName("actionObject_3")
-        self.actionModel_3 = QtWidgets.QAction(MainWindow)
-        self.actionModel_3.setObjectName("actionModel_3")
-        self.actionScene_POV = QtWidgets.QAction(MainWindow)
-        self.actionScene_POV.setObjectName("actionScene_POV")
-        self.actionAbout = QtWidgets.QAction(MainWindow)
-        self.actionAbout.setObjectName("actionAbout")
-        self.menuLoad.addAction(self.actionObject)
-        self.menuLoad.addAction(self.actionModel)
-        self.menuSave.addAction(self.actionObject_2)
-        self.menuSave.addAction(self.actionModel_2)
-        self.menuFile.addAction(self.menuLoad.menuAction())
-        self.menuFile.addAction(self.menuSave.menuAction())
-        self.menuFile.addAction(self.actionExit)
-        self.menuEdit.addAction(self.actionObject_3)
-        self.menuEdit.addAction(self.actionModel_3)
-        self.menuEdit.addAction(self.actionScene_POV)
-        self.menuTools.addAction(self.actionPredefined_Forms)
-        self.menuTools.addAction(self.actionPredefined_Scenes)
-        self.menuHelp.addAction(self.actionAbout)
-        self.menuBar.addAction(self.menuFile.menuAction())
-        self.menuBar.addAction(self.menuEdit.menuAction())
-        self.menuBar.addAction(self.menuTools.menuAction())
-        self.menuBar.addAction(self.menuHelp.menuAction())
-
-        self.retranslateUi(MainWindow)
-        QtCore.QMetaObject.connectSlotsByName(MainWindow)
-
-    def retranslateUi(self, MainWindow):
-        _translate = QtCore.QCoreApplication.translate
-        MainWindow.setWindowTitle(_translate("MainWindow", "G2-F"))
-        self.label_4.setText(_translate("MainWindow", "Object on Scene"))
-        self.label.setText(_translate("MainWindow", "Rotate   X"))
-        self.label_2.setText(_translate("MainWindow", "Rotate   Y"))
-        self.label_3.setText(_translate("MainWindow", "Rotate   Z"))
-        self.pushButton.setText(_translate("MainWindow", "Hide"))
-        self.pushButton_2.setText(_translate("MainWindow", "Modify"))
-        self.label_5.setText(_translate("MainWindow", "Zoom Ratio"))
-        self.menuFile.setTitle(_translate("MainWindow", "File"))
-        self.menuLoad.setTitle(_translate("MainWindow", "Load"))
-        self.menuSave.setTitle(_translate("MainWindow", "Save"))
-        self.menuEdit.setTitle(_translate("MainWindow", "Edit"))
-        self.menuTools.setTitle(_translate("MainWindow", "Tools"))
-        self.menuHelp.setTitle(_translate("MainWindow", "Help"))
-        self.actionExit.setText(_translate("MainWindow", "Exit"))
-        self.actionObject.setText(_translate("MainWindow", "Object"))
-        self.actionModel.setText(_translate("MainWindow", "Model"))
-        self.actionObject_2.setText(_translate("MainWindow", "Object"))
-        self.actionModel_2.setText(_translate("MainWindow", "Model"))
-        self.actionPredefined_Forms.setText(_translate("MainWindow", "Predefined Objects"))
-        self.actionPredefined_Scenes.setText(_translate("MainWindow", "Predefined Scenes"))
-        self.actionObject_3.setText(_translate("MainWindow", "Object"))
-        self.actionModel_3.setText(_translate("MainWindow", "Model"))
-        self.actionScene_POV.setText(_translate("MainWindow", "Scene POV"))
-        self.actionAbout.setText(_translate("MainWindow", "About"))
-
-
-if __name__ == "__main__":
-    import sys
-    app = QtWidgets.QApplication(sys.argv)
-    MainWindow = QtWidgets.QMainWindow()
-    ui = Ui_MainWindow()
-    ui.setupUi(MainWindow)
-    MainWindow.show()
-    sys.exit(app.exec_())
-
