--- G2-F/tetraedro.tpg
@@ -1,34 +1,0 @@
-#!OBJ
-
-N#Tetraedro
-
-E#1
-C	1.0	1.0	1.0
-P	4.3	4.3	4.3
-P	0.0	0.0	0.0
-
-E#2
-C	1.0	1.0	1.0
-P	4.3	4.3	4.3
-P	0.0	4.3	0.0
-
-E#3
-C	1.0	1.0	1.0
-P	4.3	4.3	4.3
-P	0.0	4.3	4.3
-
-E#4
-C	1.0	1.0	1.0
-P	0.0	0.0	0.0
-P	0.0	4.3	0.0
-
-E#5
-C	1.0	1.0	1.0
-P	0.0	0.0	0.0
-P	0.0	4.3	4.3
-
-E#6
-C	1.0	1.0	1.0
-P	0.0	4.3	0.0
-P	0.0	4.3	4.3
-
