+++ Proyecto-HeadMorphin/PersonalModels/3dmorph.mtl
@@ -1,0 +1,12 @@
+# Blender MTL File: '3dmorph.blend'
+# Material Count: 1
+
+newmtl defaultMat
+Ns 96.078431
+Ka 1.000000 1.000000 1.000000
+Kd 0.640000 0.640000 0.640000
+Ks 0.500000 0.500000 0.500000
+Ke 0.000000 0.000000 0.000000
+Ni 1.000000
+d 1.000000
+illum 2
