+++ libg2f/tetraedro.tpg
@@ -1,0 +1,34 @@
+#!OBJ
+
+N#Tetraedro
+
+E#1
+C	1.0	1.0	1.0
+P1	4.3	4.3	4.3
+P2	0.0	0.0	0.0
+
+E#2
+C	1.0	1.0	1.0
+P1	4.3	4.3	4.3
+P2	0.0	4.3	0.0
+
+E#3
+C	1.0	1.0	1.0
+P1	4.3	4.3	4.3
+P2	0.0	4.3	4.3
+
+E#4
+C	1.0	1.0	1.0
+P1	0.0	0.0	0.0
+P2	0.0	4.3	0.0
+
+E#5
+C	1.0	1.0	1.0
+P1	0.0	0.0	0.0
+P2	0.0	4.3	4.3
+
+E#6
+C	1.0	1.0	1.0
+P1	0.0	4.3	0.0
+P2	0.0	4.3	4.3
+
