--- prototype/README.md
@@ -1,37 +1,0 @@
-# Git-Grafica_2
-
-##Las Librerias
-
-* cuadrics
->> Contiene las 6 cuadricas Basicas y sus subtipos (12 en total)
-
-* graph_data_struct
->> Contiene las clases Color, GraphObject , Point y Edge
-la clase GraphObject crear un objeto con los metodos necesarios
-para rotar, trasladarse, ampliar, repintar y añadir texturas cara
-por cara
- 
-* grid
->> Libreria Simple: Solo contiene la clase grid, que genera
-una malla en 3D
-
-* LPlotter
->> Graficador de Palabras de la Gramatica Lindmayer 
-
-* bresenham3d
->> Implementacion del Algoritmo de Bressengam para rectas
-en el espacio
-
-* grammar_interpreter
->> Derivador Gramatical de palabras de la Gramatica Lindmayer
-la clase Grammar debe ser instanciada primero, ya que esta se
-comporta como un automata para la gramatica Lindmayer
-
-* quaternion
->> Clase Quaternion y funcion Rotate, la clase cuaternion
-ofrece todo lo necesario para instanciar un objeto tipo
-cuaterion y realizar todas las operaciones intrisecas a el,
-incluido el producto con otro cuaternion.
-
->>La Funcion Rotate emplea cuaterniones en forma matricial,
-una forma mas ligera de uso de cuaterniones pero menos comprensiva
